Reorganize project to restore online buildpack functionality
In #99 we reorganized the buildpack to be able to be packaged by
buildpack-packager. Not realizing that the online buildpack tests were commented
out, we broke the online buildpack functionality with this change.

To restore the online buildpack functionality but keep the benefits of using
buildpack-packager, the buildpack compile and supply scripts were moved back
to bin/ and the Go install script was maintained for the online buildpack
version of this product. The packager manifest had to be updated with this in
mind, and the packager script similarly had to be updated to properly prepare
the Docker environment.

Now the buildpack scripts are in the project root bin/ folder, and all other
buildpack supporting scripts are in the buildpack/ dir whether you are using
the online or offline buildpack. We may yet revisit this at a later time, but
these changes are enough to restore the online buildpack functionality for now.