Move CI scripts back into ci folder
In a previous commit, we moved them into `bin/`. But the `bin/` directory has
a special significance for buildpacks; when using this project as an online
buildpack, Cloud Foundry expects your buildpack scripts to be in that dir.

So for our test / dev utils we'll keep them in the `ci/` dir, and reserve
`bin/` for buildpack scripts.