Removes support for Conjur Enterprise v4
This change adds a `Removed` entry to the CHANGELOG.md to indicate
that support for using the buildpack with Conjur v4 is no longer
available.

No code changes are required to remove Conjur v4 support since the
code and tests are largely agnostic of which Conjur version it is
working with. The conjur-dev code and the tests use:
- cyberark/conjur latest Docker image
- cyberark/conjur-api-go v0.6.0
- cyberark/summon v0.8.2 (secretsyml module)

There is one location where the credentials for the Conjur API is
set for Conjur v5:
tests/integration/features/step_definitions/common_steps.rb, line 59.